Presenter переделан во ViewModel
