+++ app/src/main/java/ru/geekbrains/registration/mvp/ui/login/RegistrationContracts.kt
@@ -23,7 +23,7 @@ interface RegistrationContracts {
 //        fun getHandler(): Handler
 //    }
 
-    interface Presenter {
+    interface ViewModel {
         val shouldShowProgress: Signer<Boolean>
         val isSuccess: Signer<Boolean>
         val error: Signer<String>
